requirements.txt
requests
stellar-base-sseclient
SPARQLWrapper